+++ api/apis.ts
@@ -46,8 +46,8 @@ import {AxiosRequestConfig, AxiosResponse, AxiosHeaders, AxiosError, isAxiosErro
 
 export { RequestFile } from '../model/models';
 
-const revision =  "2025-07-15";
-const userAgent = "klaviyo-api-node/19.0.2";
+const revision =  "2025-10-15";
+const userAgent = "klaviyo-api-node/20.0.0";
 
 export class RetryWithExponentialBackoff {