+++ model/hasPushMarketingConsent.ts
@@ -14,7 +14,7 @@ import { HasPushMarketing } from './hasPushMarketing';
 import { PushEnum } from './pushEnum';
 export class HasPushMarketingConsent {
     'channel': PushEnum | 'push';
-    'canReceiveMarketing': HasPushMarketingConsent.CanReceiveMarketingEnum | true;
+    'canReceiveMarketing': boolean;
     'consentStatus': HasPushMarketing;
 
 
@@ -27,7 +27,7 @@ export class HasPushMarketingConsent {
         {
             "name": "canReceiveMarketing",
             "baseName": "can_receive_marketing",
-            "type": "HasPushMarketingConsent.CanReceiveMarketingEnum"
+            "type": "boolean"
         },
         {
             "name": "consentStatus",
@@ -41,7 +41,4 @@ export class HasPushMarketingConsent {
 }
 
 export namespace HasPushMarketingConsent {
-    export enum CanReceiveMarketingEnum {
-        True = <any> true
-    }
 }